% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_core_variables_are_in.r
\name{all_core_variables_are_in}
\alias{all_core_variables_are_in}
\title{(Internal) Check that data has all core variables}
\usage{
all_core_variables_are_in(mtus_data)
}
\arguments{
\item{mtus_data}{User-loaded IPUMS MTUS data which has been loaded using the \code{mtus_init} function.}
}
\value{
\code{TRUE} if all those variables are present. \code{FALSE} otherwise.
}
\description{
Examines whether the user-loaded data includes \code{AGE}, \code{TIME}, \code{ELOC}, \code{ACT_WORK}, and \code{MAIN.} These are the minimum variables needed for other data manipulation functions in the \code{gsdmtus} package.
}
