% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_valid_country.r
\name{check_valid_country}
\alias{check_valid_country}
\title{(Internal) Check that country-year mix is an available sample}
\usage{
check_valid_country(country_code)
}
\arguments{
\item{country_code}{The desired country-year code under the \code{SAMPLE} column in the MTUS data, such as "FR2009" or "UK2014".}
}
\value{
\code{TRUE} if the country-year sample is available.

\code{FALSE} if the sample isn't in the selected data; if so, the package displays a list of available countries.
}
\description{
The \code{SAMPLE} column of IPUMS MTUS data lists all country-year samples that are available. This function confirms that the user has entered an available country-year sample in their loaded IPUMS MTUS data.
}
