% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_trips.r
\name{count_trips}
\alias{count_trips}
\title{(Internal) Count the number of trips by category}
\usage{
count_trips(by_activity_data, by_person_data)
}
\arguments{
\item{by_activity_data}{By-activity data object that user has already loaded.}

\item{by_person_data}{By-person data object that user has already loaded.}
}
\value{
Newly created columns include the following: \code{work_time}, \code{home_work_time}, \code{travel_time}, \link{count of} \code{trips}, \code{commutes}, \code{nonwork_trips}, \code{pct_home_work}, \code{all_home} \link{i.e., work entirely from home}, \code{all_work}, \code{part_home}, \code{work_cat}, and \code{full_day.}
}
\description{
Returns a \code{clean_data} object with columns counting different types of trips.
}
