% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def_survey_objects.r
\name{def_survey_objects}
\alias{def_survey_objects}
\title{Define survey objects for analysis and graphing}
\usage{
def_survey_objects(workers)
}
\arguments{
\item{workers}{By-person data that the user has already loaded.}
}
\value{
\code{svy_workers}: All individuals who work.

\code{svy_all_home}: Workers who work entirely from home.

\code{svy_all_work}: Workers who work entirely away from home.

\code{svy_part_home}: Workers who work partly from home.
}
\description{
This procedure prepares four preset survey objects that are useful for exploratory data analysis and incorporate survey weights. Rather than returning values, this procedure makes four global survey-design objects available and ready for actions such as plotting graphs.
}
