% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_by_activity.r
\name{init_by_activity}
\alias{init_by_activity}
\title{(Internal) Load by-activity data}
\usage{
init_by_activity(by_activity_data, country_code)
}
\arguments{
\item{by_activity_data}{The by-activity data that the user loaded earlier using \code{mtus_init_raw_data}.}

\item{country_code}{A country-year sample available within the data. Because this involves manually typing a code (e.g. "UK2014"), the user is instead encouraged to use \code{mtus_init_countries} to avoid entry errors.}
}
\value{
\code{country_by_activity}: the by-activity data filtered to a specific country-year sample
}
\description{
This function is not meant to be user-facing. It is called within the user-facing \code{mtus_init_countries} function to create by-activity data objects; in \code{mtus_init_countries}, by-activity objects automatically start with the prefix \code{a_}, such as \code{a_FR2009}. This is distinct from \code{FR2009}, which is the convention used for cleaned data joining by-activity and by-person data.
}
\details{
The function first creates a character list of all available country-year samples by fetching all unique values from the \code{SAMPLE} column of the data set. It then runs \verb{check_valid_country(…)} to ensure that the \code{country_code} argument is valid. If so, then it calls filters \code{by_activity_data} to just
}
