% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plots.r
\name{mtus_density}
\alias{mtus_density}
\title{Example density plot using the \code{gsdmtus} package}
\usage{
mtus_density(country_data, new_plot = TRUE)
}
\description{
Example density plot using the \code{gsdmtus} package
}
