% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plots.r
\name{mtus_density_ov}
\alias{mtus_density_ov}
\title{Overlay function for example density plot using the \code{gsdmtus} package}
\usage{
mtus_density_ov(country1, country2)
}
\description{
Overlay function for example density plot using the \code{gsdmtus} package
}
