% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plots.r
\name{mtus_histogram}
\alias{mtus_histogram}
\title{Example histogram using the \code{gsdmtus} package}
\usage{
mtus_histogram(country_data)
}
\description{
Example histogram using the \code{gsdmtus} package
}
