% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtus_init_countries.r
\name{mtus_init_countries}
\alias{mtus_init_countries}
\title{Load country-specific data from IPUMS MTUS data set}
\usage{
mtus_init_countries(
  by_activity_data,
  by_person_data,
  countries,
  recode_activities = FALSE,
  recode_locations = FALSE
)
}
\arguments{
\item{by_activity_data}{By-activity data downloaded from the IPUMS website and then loaded into R with \code{mtus_init_raw_data}. The user must at minimum provide by-activity data; by-person data}

\item{by_person_data}{One of two options: either by-person data downloaded from the IPUMS website and loaded into R with \code{mtus_init_raw_data}, or \code{NULL} if the user's intent is to create standalone by-activity data objects.}

\item{countries}{The user may enter a specific country-year sample (e.g. "CA2010"), a character list of country-year samples (e.g. "FR2009" or "UK2014"), or \code{"all"}.}

\item{recode_locations}{Boolean. \code{TRUE} to recode numeric activity codes in the \code{ELOC} column with human-readable labels. \code{FALSE} to skip.}

\item{recode_activites}{Boolean. \code{TRUE} to recode numeric activity codes in the \code{MAIN} and \code{SEC} columns with human-readable labels. \code{FALSE} to skip.}
}
\value{
This procedure creates individual data objects for any given country sample (or all country samples). For instance, if a user has a large data set and only wishes to load France and UK, they can provide the character list \code{c("FR2009", "UK2014")} as the \code{countries} argument.

If the user wants to pick out every country in the data and create a separate data object for each one, the user should enter \code{"all"} as the \code{countries} argument.

To create a data object of solely by-activity data without counting numbers of trips that an individual took, set the \code{by_person_data} argument to \code{NULL}.
}
\description{
This procedure makes it easy for the user to split imported MTUS data into separate countries. The user must first use the \code{mtus_init_raw_data} function to load by-activity data and by-person data before using \code{mtus_init_countries}.
}
\details{
Please note that this procedure does not return a value; the user should call it without the assignment operator (\verb{<-}).
}
