% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtus_init_countries.r
\name{mtus_load_country}
\alias{mtus_load_country}
\title{(Internal) Load a specific country for internal use.}
\usage{
mtus_load_country(by_activity_data, by_person_data, country_code)
}
\description{
This function is not meant to be user-facing. It is called from within \code{mtus_init_countries} to count the number of trips each individual has performed. However, the user may call it (using the assignment operator \verb{<-}) as a shortcut to filter the loaded MTUS data to just a single country-year sample, join the data, and count the number of trips. The user is encouraged to use the \verb{mtus_init_countries(…)} procedure instead, though, because it is cleaner and less prone to user entry error.
}
