% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtus_recode.r
\name{mtus_recode}
\alias{mtus_recode}
\title{Recode MTUS data into human-readable labels}
\usage{
mtus_recode(input_data, which_column, codes_csv)
}
\arguments{
\item{input_data}{The by-activity data object that the user has already loaded. This argument may be entered either as a string inq uotes or as a data object without quotes.}

\item{which_column}{String denoting the column that the user wishes to recode.}

\item{codes_csv}{Either a data object or a string with path to a csv file. The file/table must contain two columns: the first column header must match the name of the MTUS variable to be renamed (e.g. MAIN, SEC, ELOC), and the second column header must be named LABEL, and the column must contain the list of corresponding human-readable labels.}
}
\value{
Returns a version of input_data with user-designated labels replacing original IPUMS MTUS codes.
}
\description{
Most IPUMS data uses numeric levels to indicate different values for variables (e.g. the number \code{8} in the \code{ELOC} column means an activity was performed at home). This function enables a user to recode numeric levels into human-readable labels. Please note that the user-inputted CSV file must have a specific format described below.
}
