% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plots.r
\name{mtus_scatter}
\alias{mtus_scatter}
\title{Example scatterplot using the \code{gsdmtus} package}
\usage{
mtus_scatter(country_data)
}
\description{
Example scatterplot using the \code{gsdmtus} package
}
