% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_plots.r
\name{mtus_trip_density}
\alias{mtus_trip_density}
\title{Example trip density plot using the \code{gsdmtus} package}
\usage{
mtus_trip_density(country_data)
}
\description{
Example trip density plot using the \code{gsdmtus} package
}
