% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recode_internal.r
\name{recode_internal}
\alias{recode_internal}
\title{(Internal) Recode activites and/or locations}
\usage{
recode_internal(input_data, activities = FALSE, locations = FALSE)
}
\arguments{
\item{input_data}{The by-activity data to recode, which may be entered either as a string inq uotes or as a data object without quotes.}

\item{locations}{Boolean; toggle whether to recode the locations in the \code{ELOC} column.}

\item{activites}{Boolean; toggle whether to recode the activities in the \code{MAIN} and \code{SEC} columns.}
}
\description{
This is an internal function that uses preset codes and labels for activites (both MAIN and SEC) and location (ELOC) variables. It is called from within \code{mtus_init_countries}. This function uses pre-loaded CSV files to run calls of the \code{mtus_recode} function.
}
